Created the "GroupReport.tex" file for the group report to be written in.
Added many comments throughout World.hs, Adventure.hs and Actions.hs in order to make the data flow of the game more evident, and also to highlight various aspects of the functionality and make the code more readable.

Created the "safeHead" and  "opposite" functions in Actions.hs, but am yet to refactor to include either. The "opposite" function may prove to be redundant, but I imagine that a future refactoring to include uses of "safeHead" may take place.