Rearranged the structure of World.hs to improve readability.
Defined the Command, Direction and Action data types in World.hs, but Command and Action will remain commented-out for the meantime until a discussion can occur in-person about implementation strategy for the grammar.

The "opposite" method was refactored to include the new Direction type. Maybe has now been removed from this method for now.

Modified "addInv" to replace if-else statement with guarded expression.