The code is non-functional in its current state and this commit is simply to backup progress.
I'm probably halfway through the refactor, only 30 errors to go!

Action and Command the type aliases have been replaced with Action and Command the concrete data types. This should help with parsing later on when we reach that stage. I attempted to take advantage of the fact that Algebraic Data Types can be recursive so that they could be used to define some sort of grammar within the program, but this ultimately concluded in a failure of epic proportion. This will be left to the Parsers once implemented later on.

The Direction data type has been fully implemented I believe. It was relatively simple and most of the work had already been done before the main refactor for Command and Action began.

None of the custom data types being defined have a Nothing value. This was considered, but ultimately seemed pointless as the Maybe type will already supply us a Nothing value if necessary.

During the refactor process, many of the variable names in the functions in Actions.hs have been modified to be more verbose, potentially against Haskell common convention. This was mostly to help improve readability, and may be undone once the refactor is complete should it be decided that this would be more idiomatic.

All instances of "head" still need to be replaced, but this is a very low priority task, as other safeguards ensure that it can't actually be used on empty lists, avoiding crashes.

Other changes include whitespace modifications for readability and some experimental yet nonfunctional code remaining in the files in a commented-out state. These instances will be discussed, but regardless of their inclusion (or lack thereof) in the final submission they will all eventually be removed.

I am currently debating as to whether Object names should be their own type (like RoomType in the current state of the program), use separate constuctors and utilise pattern matching in the Action functions, or whether they should just stick with String identifiers. The latter seems most likely.