2 errors remaining.
Most of the refactor completed. Pretty much everything in Actions.hs should be good to go, except it is not compatible with the "process" function in Adventure.hs yet because the values from the Command and Action types do not have an associated partial application that can be returned. Will attempt to resolve this issue.