1 significant error still remains.
I have undone many changes as it appears that they could make life more difficult and decrease the readability of the code. This was due to a discussion between my group members and I. I have incorporated their feedback into the design of the custom types, as they made some very good points about improving the code.

Action and Command have been restored as type alisaes, but a new Instruction data type has been introduced which still allows us to meet the custom types requirement.

The only significant error remaining is in the "process" function in Adventure.hs. Currently there is no way for the arguments to be converted into their appropriate Instruction type and argument (if applicable), so I will implement this soon as it is a critical component in the program.